requirements.txt
beautifulsoup4==4.9.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.1
Flask==1.1.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
Pillow==7.2.0
playsound==1.2.2
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.50.0
urllib3==1.25.10
Werkzeug==1.0.1
